Mesure d'un paragraphe : répondre aux propositions extrêmes de SwiftUI
Un UITextView interrogé sans largeur mesure une ligne unique et interminable, ce
qui ferait croire à une pile horizontale qu'un paragraphe veut trois mètres de
large. On répond par la largeur d'une colonne de fiche, la seule qu'un paragraphe
occupe jamais. Au passage, la largeur d'une barre de graphe est convertie
explicitement en CGFloat.

+++ Micabo/Features/Course/SheetBlockView.swift
@@ -298,7 +298,7 @@ struct SheetChartView: View {
 
                 Capsule()
                     .fill(tint.opacity(0.85))
-                    .frame(width: max(5, proxy.size.width * ratio(of: value)))
+                    .frame(width: max(5, proxy.size.width * CGFloat(ratio(of: value))))
             }
         }
         .frame(height: 9)
